overview_of_beeai/pyproject.toml
[project]
name = "overview_of_beeai"
version = "0.1.0"
description = "Overview of BeeAI workshop"
readme = "README.md"
requires-python = ">=3.13, <3.14"
dependencies = [
    "beeai-framework~=0.1.49",
    "beeai-framework[a2a]~=0.1.49",
    "beeai-sdk~=0.3.4",
    "python-dotenv~=1.1.1",
    "pydantic-ai~=0.7.6",
]
